+++ sphinx_gallery/notebook.py
@@ -113,7 +113,8 @@ def jupyter_notebook(script_blocks, gallery_conf):
     """
     first_cell = gallery_conf.get("first_notebook_cell", "%matplotlib inline")
     work_notebook = jupyter_notebook_skeleton()
-    add_code_cell(work_notebook, first_cell)
+    if first_cell is not None:
+        add_code_cell(work_notebook, first_cell)
     fill_notebook(work_notebook, script_blocks)
 
     return work_notebook